GH-187 Add --use-tls option
This commit adds the `--use-tls` option to the base CLI and passes it
through to the MinKNOW connection manager in `ru_gen.py` and
`unblock_all.py`. This is for compatibility with newer MinKNOW versions
that require/prefer secure connections.

There maybe some issues with connections that require passing an
environment variable so that the MinKNOW API can find it's certificates,
see: https://github.com/nanoporetech/minknow_api/issues/16