Include sequencing phase and guppy connection checks
This commit introduces into the dev_staging branch the ability to use
ipc and tcp guppy connections. To do this we only check the start of the
host provided in the sequencing TOML. If the string starts with 'ipc'
then an ipc style host is used: "{host}/{port}"; otherwise a tcp style
is used: "{host}:{port}".

In addition, this adds a check for sequencing phases so that the read
until loop can be paused during mux scans.