Add unblock all script
Add an unblock all script that uses a TOML file and the central
`targets.py` entry point. This has been tested using playback on a mac
but nowhere else currently. The TOML file is created as a temporary file
that is then passed through using the command line arguments.

This adds the `no_op` plugins. This is particularly useful for alignment
not so much for basecalling. It allows a hypothetical user to place all
of their logic in their basecaller module and exclude the separate
alignment step all together.