Add `readfish[mappy,guppy]` as default install target
Adding a default install target allows people to pip install readfish and get `mappy` and `ont-pyguppy-client-lib`.

Whilst the version of `ont-pyguppy-client-lib` will be likely wrong (unless on the latest version), I _think_ this is easier than people having to type `pip install readfish[X]`.

This still allows people to install using `pip install readfish[all]` with no issues. It DOES however mean that it is not possible to install without getting `mappy and guppy`, which may be an issue on some python versions. Currently however there is no way of running readfish without these so 🤷🏼‍♂️