Feature/describe plugin
> "Say what you mean and mean what you say." - Dr. Seuss

This PR adds a describe method to the ABC for plugins which implement a description of the Caller or Aligner to help the user understand what is going on.

The aligner raises a SystemExit if a contig is given as a target which is not present in the reference.

Adds a describe experiment to the Conf class, telling users what the conf sets them up to do.

* Add describe method to abc.py
Implement on no_op plugin

* Add functions used by describe methods
Mostly used by mappy.py describe
targets.py add call to describe for both plugins

* Implement describe for mappy plugin

* Add describe experiment function
Only draws regions out at the moment
Remove old describe experiment

* Send minknow messages of plugin describes
Add call to describe experiment on conf
Remove region printing from mappy plugin

* Change to simple ASCII for printing out flowcell layout
Remove ANSI codes, use `.` and `#`

* Update draw_flowcell_text to represent what symbols mean

* Fix to draw flowcell split using actual channels in Region
Now takes all arguments that are passed to generate_flowcell
Works on prom and gridion

* Update Conf and get_coords docstring
describe _channel_map member on Conf
Add return type to get_coords

* Add barcode experiment description

* Add barcode region description

* Fix to _check_inf
IF we have no targets or target is not on contig return 0

* Add tests for the describe functionality

* Raise SystemExit on target contig not being reference
prints out the missing contigs

* Fix tests to include catching systemexit for missing contigs
And patch the expected messaged for tests with correct contigs

* Remove new line, from missing contigs error message

* Describe the toml when validating it.
Adds a --no-describe option to skip description with validation.
This is set in the validation test. if there are errors validating, we skip description,
as it probably will not work.

* Move conditional error check for validation describe
Typo'd it into the wrong check

* Update documentation for validate
Describe `--no-describe` flag.

* Add describe function into developer guide

* DRYer tests
Use parameterise to make tests a lot less repetitive
Alter fh.read() text based on which aligner is available in the environment

* Addressing final comments