Implement runsafe and additional logging.
Check whether the first read readfish sees has begun before readfish connected, outside of mux scan. If this is the case it is always sequenced - as we have no idea how long it is.

* Implementing first read checking, a last action counter and a dry-run option

* Implementing message if ignoring first read.

* Correctly implementing attrs and adding docstrings.

* Changing runsafe to started_before_sequencing. Moving last_sent_action and started_before_sequencing to init. Logging action override.

* Resolving dry-run implementation.

* Adding doctests for Action and ActionTracker
Renaming LastReadSeen class and methods for clarity - but I think this is good

* Address comments

* Reversed logic on which phase we started in
This way we aren't doing a load of checks against `not self.started_in....`
Simplifies reading through ( _I think_ )
Tidied up some comments

* Replacing `__attrs_post_init__` with `attrs.Factory`
PreviouslySentActionTracker now defaults to a dict, rather than
overwrites with one.

---------

Co-authored-by: Adoni5 <roryjmunro1@gmail.com>